Cleanup, structure and document.
